extracted notebook policy helpers to own file
