+++ backend/src/ml_space_lambda/app_configuration/lambda_functions.py
@@ -18,15 +18,10 @@ import json
 import logging
 import time
 
-import boto3
-
+from ml_space_lambda.app_configuration.policy_helper.notebook import update_instance_constraint_policies
 from ml_space_lambda.data_access_objects.app_configuration import AppConfigurationDAO, AppConfigurationModel, SettingsModel
-from ml_space_lambda.enums import ResourceType
 from ml_space_lambda.utils.common_functions import api_wrapper
-from ml_space_lambda.utils.iam_manager import IAMManager
-from ml_space_lambda.utils.mlspace_config import get_environment_variables, retry_config
 
-iam = boto3.client("iam", config=retry_config)
 log = logging.getLogger(__name__)
 app_configuration_dao = AppConfigurationDAO()
 
@@ -46,13 +41,7 @@ def update_configuration(event, context):
     version_id = request["versionId"] + 1  # increment so this will be the latest version
 
     new_configuration = SettingsModel.from_dict(request["configuration"])
-
-    previous_app_configurations = app_configuration_dao.get(configScope=configScope, num_versions=1)
-    if len(previous_app_configurations) > 0:
-        previous_configuration = previous_app_configurations[0].configuration
-        update_instance_constraint_policies(
-            previous_configuration.disabled_instance_types, new_configuration.disabled_instance_types, context
-        )
+    update_instance_constraint_policies(new_configuration.enabled_instance_types, context)
 
     app_configuration = AppConfigurationModel(
         configScope=configScope,
@@ -72,66 +61,3 @@ def update_configuration(event, context):
         raise e
 
     return f"Successfully updated configuration for {configScope}, version {version_id}."
-
-
-def create_instance_constraint_statement(actions: list[str], resources: list[str], allowed_instances: list[str]):
-    return {
-        "Effect": "Allow",
-        "Action": actions,
-        "Resource": resources,
-        "Condition": {"ForAnyValue:StringEquals": {"sagemaker:InstanceTypes": allowed_instances}},
-    }
-
-
-def create_sagemaker_resource_arn(resource: str, context) -> str:
-    components = context.invoked_function_arn.split(":")
-    return f"arn:{components[1]}:sagemaker:{components[3]}:{components[4]}:{resource}/*"
-
-
-def update_instance_constraint_policies(previous_configuration, new_configuration, context) -> None:
-    env_vars = get_environment_variables()
-
-    # update jobs
-    actions = ["sagemaker:CreateTrainingJob", "sagemaker:CreateHyperParameterTuningJob", "sagemaker:CreateTransformJob"]
-    resources = [
-        create_sagemaker_resource_arn(ResourceType.TRAINING_JOB.value, context),
-        create_sagemaker_resource_arn(ResourceType.HPO_JOB.value, context),
-        create_sagemaker_resource_arn(ResourceType.TRANSFORM_JOB.value, context),
-    ]
-
-    training_statement = create_instance_constraint_statement(
-        actions,
-        resources,
-        new_configuration.training_job_instance_types,
-    )
-
-    transform_statement = create_instance_constraint_statement(
-        actions, resources, new_configuration.transform_jobs_instance_types
-    )
-
-    create_instance_constraint_policy_version(
-        env_vars["JOB_INSTANCE_CONSTRAINT_POLICY_ARN"], [training_statement, transform_statement]
-    )
-
-    # endpoint config
-    actions = ["sagemaker:CreateEndpointConfig"]
-    resources = [create_sagemaker_resource_arn(ResourceType.ENDPOINT.value, context)]
-    endpoint_statement = create_instance_constraint_statement(actions, resources, new_configuration.endpoint_instance_types)
-    create_instance_constraint_policy_version(env_vars["ENDPOINT_CONFIG_INSTANCE_CONSTRAINT_POLICY_ARN"], [endpoint_statement])
-
-
-def create_instance_constraint_policy_version(policy_arn: str, statements: list) -> None:
-    log.info("Creating new version %s", policy_arn)
-    iam.create_policy_version(
-        PolicyArn=policy_arn,
-        PolicyDocument=json.dumps(
-            {
-                "Version": "2012-10-17",
-                "Statement": statements,
-            }
-        ),
-        SetAsDefault=True,
-    )
-
-    iam_manager = IAMManager(iam)
-    iam_manager._delete_unused_policy_versions(policy_arn)
